Move get_spectrum_name function to utilities module.
